chore!: Remove unmaintained typescript-require hook
