+++ test/fixtures/ts/2/.babelrc
