+++ test/fixtures/ts/2/component.ts
