+++ test/fixtures/ts/2/test.ts
@@ -1,25 +1,5 @@
-var test: {
-  data: {
-    trueKey: boolean,
-    falseKey: boolean,
-    subKey: {
-      subProp: number
-    }
-  }
-} = {
-  data: {
-    trueKey: true,
-    falseKey: false,
-    subKey: {
-      subProp: 1
-    }
-  }
-};
+import Component from "./component"
 
-var main: {
-  default: typeof test
-} = {
-  default: test
-};
+var test: typeof Component = { ...Component }
 
-export = main;
+export default test
