--- test/fixtures/ts/2/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "declaration": false,
-    "noImplicitAny": false,
-    "removeComments": true,
-    "sourceMap": true,
-    "outDir": ".tmp"
-  }
-}
