+++ test/fixtures/ts/4/.swcrc
